requirements.txt
flask
requests
pycountry
gunicorn
python-dotenv